requirements.txt
aiomysql==0.2.0
channels==3.0.1
Django==4.1
djangorestframework==3.14.0
joblib==1.3.2
mysql_connector_repackaged==0.3.1
numpy==1.26.2
pandas==2.1.4
PyMySQL==1.1.0
simplejson==3.19.2
termcolor==2.4.0
torch==2.1.1
watchdog==3.0.0
websocket_client==1.7.0
websockets==12.0
